requirements.txt
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
scikit-learn==1.3.0
pytorch-metric-learning==2.3.0
fastapi==0.103.0
uvicorn==0.23.2
numpy==1.24.3